fix tests following suggestions
